# api/budget/api.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Budget Service
tags:
  - name: AccountList
paths:
  /account/{budgetID}:
    post:
      tags:
        - AccountList
      summary: Add a new account to specific list
      operationId: add
      parameters:
        - name: budgetID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                currency:
                  type: string
                initialBalance:
                  type: integer
              required:
                - title
                - currency
                - initialBalance
        required: true
      responses:
        204:
          description: 'Success result'
        400:
          description: 'Returns when invalid parameters were specified'
        409:
          description: 'Returns when the title is duplicated'
  /account/{budgetID}/{accountID}/order/{position}:
    put:
      tags:
        - AccountList
      summary: Change an order for an account to specific list
      operationId: reorder
      parameters:
        - name: budgetID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: accountID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: position
          in: path
          required: true
          schema:
            type: integer
      responses:
        204:
          description: 'Success result'
        400:
          description: 'Returns when invalid parameters were specified'
        404:
          description: 'Returns when the account is not exist or deleted'
  /account/{budgetID}/{accountID}/title:
    put:
      tags:
        - AccountList
      summary: Change a title for an account to specific list
      operationId: rename
      parameters:
        - name: budgetID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: accountID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
              required:
                - title
        required: true
      responses:
        204:
          description: 'Success result'
        400:
          description: 'Returns when invalid parameters were specified'
        404:
          description: 'Returns when the account is not exist or deleted'
        409:
          description: 'Returns when the title is duplicated'
  /account/{budgetID}/{accountID}/status/{status}:
    put:
      tags:
        - AccountList
      summary: Change a status for an account to specific list
      operationId: changeStatus
      parameters:
        - name: budgetID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: accountID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: path
          required: true
          schema:
            type: string
            enum:
              - active
              - cancelled
      responses:
        204:
          description: 'Success result'
        400:
          description: 'Returns when invalid parameters were specified'
        403:
          description: 'Returns when the status cannot be changed to specified'
        404:
          description: 'Returns when the account is not exist or deleted'
  /account/{budgetID}/{accountID}:
    delete:
      tags:
        - AccountList
      summary: Delete an account to specific list
      operationId: delete
      parameters:
        - name: budgetID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: accountID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: 'Success result'
        400:
          description: 'Returns when invalid parameters were specified'
        404:
          description: 'Returns when the account is not exist or deleted'